simplify vertex selection composition/dataflow, use new thunk within Overlay
